refactor(cli): TECHDEBT-02 — mueve build_parser/_main/main/_print_result de core.py a cli.py
core.py (god module de 2654 lineas) pierde las ~640 lineas de CLI:
build_parser, _main, main, _print_result y _output_dataframe van a cli.py
(713 lineas). core.py queda solo con la API publica (1950 lineas).

- __init__.py y el shim src/chile_hub.py re-exportan main desde cli.
- cli.py importa ChileHubDataManager desde core y ChileHubError desde
  exceptions (core no re-exporta excepciones).
- import lazy de polars en _main (validate) — antes venia del import de
  modulo de core.
- Verificado con wheel instalado: chile-hub version/list funcionan fuera
  del checkout.
